requirements.txt
fastapi
uvicorn[standard]
stackapi
azure-storage-blob
python-dotenv
gunicorn